+++ packages/fe-fpm-writer/test/unit/action.test.ts
@@ -347,6 +347,45 @@ describe('CustomAction', () => {
                     expect(fs.read(existingPath)).toMatchSnapshot();
                 }
             );
+
+            test('"eventHandler" is "object", append new function to controller extension', () => {
+                const fileName = 'MyExistingAction';
+                // Create existing file with existing actions
+                const folder = join('ext', 'fragments');
+                const existingPath = join(testDir, 'webapp', folder, `${fileName}.js`);
+                // Generate handler with single method - content should be updated during generating of custom action
+                fs.copyTpl(join(__dirname, '../../templates', 'common/EventHandler.js'), existingPath, {
+                    eventHandlerFnName: 'onPress'
+                });
+                // Create third action - append existing js file
+                const actionName = 'CustomAction2';
+                const fnName = 'onHandleSecondAction';
+                const controllerPrefix = '.extension';
+                // use controller prefix to make sure it is controller extension
+                generateCustomActionWithEventHandler(
+                    actionName,
+                    {
+                        fnName,
+                        fileName,
+                        controllerPrefix,
+                        insertScript: {
+                            fragment:
+                                ',\n        onHandleSecondAction: function() {\n            MessageToast.show("Custom handler invoked.");\n        }',
+                            position: {
+                                line: 8,
+                                character: 9
+                            }
+                        }
+                    },
+                    folder
+                );
+                const manifest = fs.readJSON(join(testDir, 'webapp', 'manifest.json')) as Manifest;
+                const action = getActionByName(manifest, actionName);
+                // Check if action press has controller prefix added
+                expect(action['press']).toEqual(`${controllerPrefix}.my.test.App.ext.fragments.${fileName}.${fnName}`);
+                // Check update js file content
+                expect(fs.read(existingPath)).toMatchSnapshot();
+            });
         });
 
         describe('Test property custom "tabSizing"', () => {